disentangle objective.py file
