+++ pypesto/objective/__init__.py
@@ -5,9 +5,9 @@ Objective
 """
 
 from .objective import (ObjectiveOptions,
-                        Objective,
-                        res_to_fval)
+                        Objective)
 from .amici_objective import AmiciObjective
+from .util import res_to_fval
 
 __all__ = ["ObjectiveOptions",
            "Objective",
